td8 : catalog impl csv in progress

+++ td/td8/src/main/java/fr/ubordeaux/ao/referencemanagement/infrastructure/csv/CatalogCSVImpl.java
@@ -1,10 +1,10 @@
 package fr.ubordeaux.ao.referencemanagement.infrastructure.csv;
 
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
@@ -26,11 +26,15 @@ public class CatalogCSVImpl implements Catalog {
             throw new ReferenceManagementException("filename is null");
         }
         try {
-            fileWriter = new FileWriter(filename);            
-            fileWriter.write("SubCatalogName;ReferenceID;ReferenceName;ReferenceDescription;ReferenceBasePrice\n");
-            fileWriter.flush();
-
-            fileReader = new BufferedReader(new FileReader(filename));            
+            File file = new File(filename);
+            fileWriter = new FileWriter(filename, true);
+            if ((file.exists() && file.length() == 0) ||
+                !file.exists()) {           
+                fileWriter.write("SubCatalogName;ReferenceID;ReferenceName;ReferenceDescription;ReferenceBasePrice\n");
+                fileWriter.flush();
+            }
+            fileReader = new BufferedReader(new FileReader(filename));
+            fileReader.readLine();            
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -42,11 +46,15 @@ public class CatalogCSVImpl implements Catalog {
             throw new ReferenceManagementException("filename is null");
         }
         try {
-            fileWriter = new FileWriter(filename);            
-            fileWriter.write("SubCatalogName;ReferenceID;ReferenceName;ReferenceDescription;ReferenceBasePrice\n");
-            fileWriter.flush();
+            File file = new File(filename);
+            if (!file.exists()) {
+                fileWriter = new FileWriter(filename);            
+                fileWriter.write("SubCatalogName;ReferenceID;ReferenceName;ReferenceDescription;ReferenceBasePrice\n");
+                fileWriter.flush();
 
-            fileReader = new BufferedReader(new FileReader(filename));            
+                fileReader = new BufferedReader(new FileReader(filename));
+                fileReader.readLine();
+            }            
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -81,20 +89,18 @@ public class CatalogCSVImpl implements Catalog {
         try {
             while ((line = fileReader.readLine()) != null) {
                 String[] lineSplit = line.split(";");
-                UUID referenceID = null;
-                for (String s : lineSplit) {
-                    if (s.equals("SubCatalogName")) {
-                        continue;
-                    }
-                    else {
-
-                    }
-                }
+                UUID referenceID = UUID.fromString(lineSplit[1]);
+                String referenceName = lineSplit[2];
+                String referenceDescription = lineSplit[3];
+                Price referenceBasePrice = new Price(Integer.parseInt(lineSplit[4]));
+                Reference reference = new Reference(referenceID, referenceName, referenceDescription, referenceBasePrice);
+                referenceSet.add(reference);
             }
+            return referenceSet;
         } catch (IOException e) {
             e.printStackTrace();
         }
-        return null;
+        return referenceSet;
     }
 
     @Override
@@ -115,7 +121,7 @@ public class CatalogCSVImpl implements Catalog {
         Price price = reference.getBasePrice();
         try {
             fileWriter.write(catalogName + ";" + referenceID.toString() + ";" + referenceName + ";"
-                    + referenceDescription + ";" + price.toString());
+                    + referenceDescription + ";" + price.toString() + "\n");
             fileWriter.flush();
         } catch (IOException e) {
             e.printStackTrace();
